+++ crates/cli/src/cmd/check.rs
@@ -6,8 +6,17 @@ pub struct Args {
     /// Path to project
     #[clap(short, long)]
     directory: Option<PathBuf>,
+
+    /// Skip tests; run only the type-checker
+    #[clap(short, long)]
+    skip_tests: bool,
 }
 
-pub fn exec(Args { directory }: Args) -> miette::Result<()> {
-    crate::with_project(directory, |p| p.check())
+pub fn exec(
+    Args {
+        directory,
+        skip_tests,
+    }: Args,
+) -> miette::Result<()> {
+    crate::with_project(directory, |p| p.check(skip_tests))
 }